+++ src/azure-cli-core/azure/cli/core/_output.py
@@ -9,12 +9,20 @@ import knack.output
 class AzOutputProducer(knack.output.OutputProducer):
     def __init__(self, cli_ctx=None):
         super(AzOutputProducer, self).__init__(cli_ctx)
-        super(AzOutputProducer, self)._FORMAT_DICT['yaml'] = self.format_yaml
+        additional_formats = {
+            'yaml': self.format_yaml,
+            'none': self.format_none
+        }
+        super(AzOutputProducer, self)._FORMAT_DICT.update(additional_formats)
 
     @staticmethod
     def format_yaml(obj):
         import yaml
         return yaml.safe_dump(obj.result, default_flow_style=False)
+
+    @staticmethod
+    def format_none(_):
+        return ""
 
     def check_valid_format_type(self, format_type):
         return format_type in self._FORMAT_DICT