+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/task.py
@@ -441,7 +441,7 @@ def acr_task_list_runs(cmd,
     if image:
         from .repository import get_image_digest
         try:
-            repository, _, manifest = get_image_digest(cmd.cli_ctx, registry_name, resource_group_name, image)
+            repository, _, manifest = get_image_digest(cmd.cli_ctx, registry_name, image)
             filter_str = _add_run_filter(
                 filter_str, 'OutputImageManifests', '{}@{}'.format(repository, manifest), 'contains')
         except CLIError as e: