+++ src/command_modules/azure-cli-acs/azure/cli/command_modules/acs/_help.py
@@ -221,6 +221,7 @@ helps['aks create'] = """
             These addons are available:
                 http_application_routing - configure ingress with automatic public DNS name creation.
                 monitoring - turn on Log Analytics monitoring. Uses the Log Analytics Default Workspace if it exists, else creates one. Specify "--workspace-resource-id" to use an existing workspace.
+                virtual-node - enable AKS Virtual Node (PREVIEW). Requires --subnet_name to provide the name of an existing subnet for the Virtual Node to use.
         - name: --disable-rbac
           type: bool
           short-summary: Disable Kubernetes Role-Based Access Control.
@@ -291,6 +292,7 @@ helps['aks enable-addons'] = """
       These addons are available:
           http_application_routing - configure ingress with automatic public DNS name creation.
           monitoring - turn on Log Analytics monitoring. Requires "--workspace-resource-id".
+          virtual-node - enable AKS Virtual Node (PREVIEW). Requires --subnet_name to provide the name of an existing subnet for the Virtual Node to use.
     parameters:
         - name: --addons -a
           type: string