+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/custom.py
@@ -906,12 +906,14 @@ def update_ddos_plan(cmd, instance, tags=None, vnets=None):
 
     if tags is not None:
         instance.tags = tags
-    if vnets == "":
-        vnets = []
     if vnets is not None:
         logger.info('Attempting to update the VNets attached to the DDoS protection plan.')
-        vnet_ids = set([x.id for x in vnets])
-        existing_vnet_ids = set([x.id for x in instance.virtual_networks or []])
+        vnet_ids = set([])
+        if len(vnets) == 1 and not vnets[0]:
+            pass
+        else:
+            vnet_ids = {x.id for x in vnets}
+        existing_vnet_ids = {x.id for x in instance.virtual_networks} or set([])
         client = network_client_factory(cmd.cli_ctx).virtual_networks
         for vnet_id in vnet_ids.difference(existing_vnet_ids):
             logger.info("Adding VNet '%s' to plan.", vnet_id)
@@ -1113,26 +1115,26 @@ def _build_record(cmd, data):
     try:
         if record_type == 'aaaa':
             return AaaaRecord(ipv6_address=data['ip'])
-        elif record_type == 'a':
+        if record_type == 'a':
             return ARecord(ipv4_address=data['ip'])
-        elif (record_type == 'caa' and
-              supported_api_version(cmd.cli_ctx, ResourceType.MGMT_NETWORK_DNS, min_api='2018-03-01-preview')):
+        if (record_type == 'caa' and
+                supported_api_version(cmd.cli_ctx, ResourceType.MGMT_NETWORK_DNS, min_api='2018-03-01-preview')):
             return CaaRecord(value=data['value'], flags=data['flags'], tag=data['tag'])
-        elif record_type == 'cname':
+        if record_type == 'cname':
             return CnameRecord(cname=data['alias'])
-        elif record_type == 'mx':
+        if record_type == 'mx':
             return MxRecord(preference=data['preference'], exchange=data['host'])
-        elif record_type == 'ns':
+        if record_type == 'ns':
             return NsRecord(nsdname=data['host'])
-        elif record_type == 'ptr':
+        if record_type == 'ptr':
             return PtrRecord(ptrdname=data['host'])
-        elif record_type == 'soa':
+        if record_type == 'soa':
             return SoaRecord(host=data['host'], email=data['email'], serial_number=data['serial'],
                              refresh_time=data['refresh'], retry_time=data['retry'], expire_time=data['expire'],
                              minimum_ttl=data['minimum'])
-        elif record_type == 'srv':
+        if record_type == 'srv':
             return SrvRecord(priority=data['priority'], weight=data['weight'], port=data['port'], target=data['target'])
-        elif record_type in ['txt', 'spf']:
+        if record_type in ['txt', 'spf']:
             text_data = data['txt']
             return TxtRecord(value=text_data) if isinstance(text_data, list) else TxtRecord(value=[text_data])
     except KeyError as ke:
@@ -3008,7 +3010,8 @@ def create_traffic_manager_profile(cmd, traffic_manager_profile_name, resource_g
                                    routing_method, unique_dns_name, monitor_path=None,
                                    monitor_port=80, monitor_protocol=MonitorProtocol.http.value,
                                    profile_status=ProfileStatus.enabled.value,
-                                   ttl=30, tags=None, interval=None, timeout=None, max_failures=None):
+                                   ttl=30, tags=None, interval=None, timeout=None, max_failures=None,
+                                   monitor_custom_headers=None, status_code_ranges=None):
     from azure.mgmt.trafficmanager import TrafficManagerManagementClient
     from azure.mgmt.trafficmanager.models import Profile, DnsConfig, MonitorConfig
     client = get_mgmt_service_client(cmd.cli_ctx, TrafficManagerManagementClient).profiles
@@ -3022,13 +3025,16 @@ def create_traffic_manager_profile(cmd, traffic_manager_profile_name, resource_g
                                                    path=monitor_path,
                                                    interval_in_seconds=interval,
                                                    timeout_in_seconds=timeout,
-                                                   tolerated_number_of_failures=max_failures))
+                                                   tolerated_number_of_failures=max_failures,
+                                                   custom_headers=monitor_custom_headers,
+                                                   expected_status_code_ranges=status_code_ranges))
     return client.create_or_update(resource_group_name, traffic_manager_profile_name, profile)
 
 
 def update_traffic_manager_profile(instance, profile_status=None, routing_method=None, tags=None,
                                    monitor_protocol=None, monitor_port=None, monitor_path=None,
-                                   ttl=None, timeout=None, interval=None, max_failures=None):
+                                   ttl=None, timeout=None, interval=None, max_failures=None,
+                                   monitor_custom_headers=None, status_code_ranges=None):
     if tags is not None:
         instance.tags = tags
     if profile_status is not None:
@@ -3052,6 +3058,10 @@ def update_traffic_manager_profile(instance, profile_status=None, routing_method
         instance.monitor_config.timeout_in_seconds = timeout
     if max_failures is not None:
         instance.monitor_config.tolerated_number_of_failures = max_failures
+    if monitor_custom_headers is not None:
+        instance.monitor_config.custom_headers = monitor_custom_headers
+    if status_code_ranges is not None:
+        instance.monitor_config.expected_status_code_ranges = status_code_ranges
 
     # TODO: Remove workaround after https://github.com/Azure/azure-rest-api-specs/issues/1940 fixed
     for endpoint in instance.endpoints:
@@ -3066,7 +3076,8 @@ def create_traffic_manager_endpoint(cmd, resource_group_name, profile_name, endp
                                     target_resource_id=None, target=None,
                                     endpoint_status=None, weight=None, priority=None,
                                     endpoint_location=None, endpoint_monitor_status=None,
-                                    min_child_endpoints=None, geo_mapping=None):
+                                    min_child_endpoints=None, geo_mapping=None,
+                                    monitor_custom_headers=None, subnets=None):
     from azure.mgmt.trafficmanager import TrafficManagerManagementClient
     from azure.mgmt.trafficmanager.models import Endpoint
     ncf = get_mgmt_service_client(cmd.cli_ctx, TrafficManagerManagementClient).endpoints
@@ -3076,7 +3087,9 @@ def create_traffic_manager_endpoint(cmd, resource_group_name, profile_name, endp
                         endpoint_location=endpoint_location,
                         endpoint_monitor_status=endpoint_monitor_status,
                         min_child_endpoints=min_child_endpoints,
-                        geo_mapping=geo_mapping)
+                        geo_mapping=geo_mapping,
+                        subnets=subnets,
+                        custom_headers=monitor_custom_headers)
 
     return ncf.create_or_update(resource_group_name, profile_name, endpoint_type, endpoint_name,
                                 endpoint)
@@ -3085,7 +3098,8 @@ def create_traffic_manager_endpoint(cmd, resource_group_name, profile_name, endp
 def update_traffic_manager_endpoint(instance, endpoint_type=None, endpoint_location=None,
                                     endpoint_status=None, endpoint_monitor_status=None,
                                     priority=None, target=None, target_resource_id=None,
-                                    weight=None, min_child_endpoints=None, geo_mapping=None):
+                                    weight=None, min_child_endpoints=None, geo_mapping=None,
+                                    subnets=None, monitor_custom_headers=None):
     if endpoint_location is not None:
         instance.endpoint_location = endpoint_location
     if endpoint_status is not None:
@@ -3104,6 +3118,10 @@ def update_traffic_manager_endpoint(instance, endpoint_type=None, endpoint_locat
         instance.min_child_endpoints = min_child_endpoints
     if geo_mapping is not None:
         instance.geo_mapping = geo_mapping
+    if subnets is not None:
+        instance.subnets = subnets
+    if monitor_custom_headers:
+        instance.custom_headers = monitor_custom_headers
 
     return instance